add XML content custom support
